requirements.txt
streamlit
langchain
langchain-core
langchain-nvidia-ai-endpoints
python-dotenv